refactor separate game context menus from ui base
