+++ frontend/src/components/(game)/(context-menus)/(menu-items)/draw-to.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@/components/(ui)/button";
-import { ContextMenuItem } from "@/components/(ui)/context-menu";
+import { ContextMenuItem } from "@/components/(game)/(context-menus)/context-menu";
 import { Input } from "@/components/(ui)/input";
 import {
     Popover,
