+++ frontend/src/components/(game)/(context-menus)/context-menu.tsx
@@ -1,0 +1,56 @@
+"use client";
+
+import * as React from "react";
+import * as ContextMenuPrimitive from "@radix-ui/react-context-menu";
+
+import { useArrowsContext } from "@/contexts/(game)/arrows";
+import {
+    ContextMenuCheckboxItem,
+    ContextMenuContent,
+    ContextMenuGroup,
+    ContextMenuItem,
+    ContextMenuLabel,
+    ContextMenuPortal,
+    ContextMenuRadioGroup,
+    ContextMenuRadioItem,
+    ContextMenuSeparator,
+    ContextMenuShortcut,
+    ContextMenuSub,
+    ContextMenuSubContent,
+    ContextMenuSubTrigger,
+} from "@/components/(ui)/context-menu";
+
+const ContextMenu = ContextMenuPrimitive.Root;
+
+const ContextMenuTrigger = React.forwardRef<
+    React.ElementRef<typeof ContextMenuPrimitive.Trigger>,
+    React.ComponentPropsWithoutRef<typeof ContextMenuPrimitive.Trigger>
+>(({ ...props }, ref) => {
+    const { isActive } = useArrowsContext();
+    return (
+        <ContextMenuPrimitive.Trigger
+            disabled={isActive}
+            ref={ref}
+            {...props}
+        />
+    );
+});
+ContextMenuTrigger.displayName = ContextMenuPrimitive.Trigger.displayName;
+
+export {
+    ContextMenu,
+    ContextMenuTrigger,
+    ContextMenuContent,
+    ContextMenuItem,
+    ContextMenuCheckboxItem,
+    ContextMenuRadioItem,
+    ContextMenuLabel,
+    ContextMenuSeparator,
+    ContextMenuShortcut,
+    ContextMenuGroup,
+    ContextMenuPortal,
+    ContextMenuSub,
+    ContextMenuSubContent,
+    ContextMenuSubTrigger,
+    ContextMenuRadioGroup,
+};
