+++ frontend/src/components/(game)/(context-menus)/search-card.tsx
@@ -8,7 +8,7 @@ import {
     ContextMenuSubContent,
     ContextMenuSubTrigger,
     ContextMenuTrigger,
-} from "@/components/(ui)/context-menu";
+} from "@/components/(game)/(context-menus)/context-menu";
 import { useShahrazadGameContext } from "@/contexts/(game)/game";
 import { ShahrazadActionCase } from "@/types/bindings/action";
 import { type ReactNode } from "react";
